// src/app/generated/prisma/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AppRole = {
  WORKER: 'WORKER',
  ADMIN: 'ADMIN'
} as const

export type AppRole = (typeof AppRole)[keyof typeof AppRole]


export const JobStatus = {
  DRAFT: 'DRAFT',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  ON_HOLD: 'ON_HOLD'
} as const

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus]


export const ProcessStatus = {
  QUEUED: 'QUEUED',
  ACTIVE: 'ACTIVE',
  COMPLETED: 'COMPLETED',
  SKIPPED: 'SKIPPED'
} as const

export type ProcessStatus = (typeof ProcessStatus)[keyof typeof ProcessStatus]